add test for call method that calling map.set/map.get on cb
